protect all menu routes and add redis cache

+++ app/menu/option/api.py
@@ -1,6 +1,9 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.auth.database import get_async_session
+from starlette import status
+
+from app.auth.auth_backend import current_active_user
+from app.auth.database import get_async_session, User
 
 from app.menu.option.crud import create_option, get_options, get_option, update_option, delete_option
 from app.menu.option.schema import OptionCreate, OptionUpdate, OptionResponse
@@ -13,8 +16,11 @@ router = APIRouter()
 @router.post("/", response_model=OptionResponse)
 async def create_option_endpoint(
         option: OptionCreate,
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
 
     return await create_option(db, option)
 
@@ -22,8 +28,11 @@ async def create_option_endpoint(
 @router.get("/", )
 @cache(expire=60)
 async def get_options_endpoint(
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await get_options(db)
 
 
@@ -31,8 +40,11 @@ async def get_options_endpoint(
 @cache(expire=60)
 async def get_option_endpoint(
         option_id: int,
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await get_option(db, option_id)
 
 
@@ -40,14 +52,20 @@ async def get_option_endpoint(
 async def update_option_endpoint(
         option_id: int,
         option: OptionUpdate,
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await update_option(db, option_id, option)
 
 
 @router.delete("/{option_id}")
 async def delete_option_endpoint(
         option_id: int,
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await delete_option(db, option_id)
