+++ app/menu/product/api.py
@@ -1,8 +1,11 @@
 from typing import Optional, Union
 
-from fastapi import APIRouter, Depends, UploadFile, Form, File
+from fastapi import APIRouter, Depends, UploadFile, Form, File, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.auth.database import get_async_session
+from starlette import status
+
+from app.auth.auth_backend import current_active_user
+from app.auth.database import get_async_session, User
 
 from app.menu.product.crud import create_product, get_products, get_product, update_product, delete_product
 from app.menu.product.schema import ProductCreate, ProductUpdate, ProductResponse
@@ -20,8 +23,12 @@ async def create_product_endpoint(
         status: bool = Form(...),
         sort_order: int = Form(...),
         image: UploadFile = File(...),
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise Exception("Only superuser can create product")
+
     product = ProductCreate(
         name=name,
         price=price,
@@ -36,14 +43,23 @@ async def create_product_endpoint(
 @router.get("/")
 @cache(expire=60)
 async def get_products_endpoint(
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await get_products(db)
 
 
 @router.get("/{product_id}")
 @cache(expire=60)
-async def get_product_endpoint(product_id: int, db: AsyncSession = Depends(get_async_session)):
+async def get_product_endpoint(
+        product_id: int,
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
+):
+    if not current_user:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized")
     return await get_product(db, product_id)
 
 
@@ -56,8 +72,11 @@ async def update_product_endpoint(
         status: Optional[bool] = Form(None),
         sort_order: Optional[int] = Form(None),
         image: Union[Optional[UploadFile], str] = File(None),
-        db: AsyncSession = Depends(get_async_session)
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
 ):
+    if not current_user.is_superuser:
+        raise Exception("Only superuser can update product")
     product = ProductUpdate(
         name=name,
         price=price,
@@ -69,5 +88,11 @@ async def update_product_endpoint(
 
 
 @router.delete("/{product_id}")
-async def delete_product_endpoint(product_id: int, db: AsyncSession = Depends(get_async_session)):
+async def delete_product_endpoint(
+        product_id: int,
+        db: AsyncSession = Depends(get_async_session),
+        current_user: User = Depends(current_active_user)
+):
+    if not current_user.is_superuser:
+        raise Exception("Only superuser can delete product")
     return await delete_product(db, product_id)
